put find_bad_apples test in its own file Closes #43
